+++ src/scalecast/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.19.3'
+__version__ = '0.19.4'
 
 from .util import metrics
 import inspect